Added some cleanup of threads and events on unmount

+++ build/fuzzysearch.js
@@ -370,6 +370,7 @@ var FuzzySearch = React.createClass({displayName: "FuzzySearch",
 
 	componentWillUnmount: function(){
 		this._closeWorkers();
+		document.removeEventListener("click", this._checkForClose)
 	},
 
 	_checkForClose: function (e) {
@@ -388,6 +389,7 @@ var FuzzySearch = React.createClass({displayName: "FuzzySearch",
 
 	_createWorkers: function(){
 		if(this.props.useWebWorkers && _canUseWorkers){
+			this._closeWorkers();
 			this._threads = [];
 			
 			var workerBlob;
@@ -434,7 +436,8 @@ var FuzzySearch = React.createClass({displayName: "FuzzySearch",
 	_closeWorkers: function(){
 		if(this.props.useWebWorkers && this._threads){
 			this._threads.forEach(function(thread){
-				thread.worker.terminate();
+				if(thread.worker && thread.worker.terminate)
+					thread.worker.terminate();
 			})
 		}
 	},
